Final Contract Updates (#92)
* add staking contract for range positions

* supportsInterface method

* format storage file

* add missing contracts

* add staking and unstaking tests

* staking tests

* validate canoncial limit pools; track total fee growth for pool

* pre testnet deployment

* historical subgraph for limit and swap

* add new files

* load feegrowth from storage

* subgraph updates for season rewards

* order history schema update

* manual gas limit for burnrangestake

* subgraph updates for mint limit event

* rangestaker test coverage

* symbol by pool pair

* subgraph updates; token symbol and name

* simplify revert cases for positions

* POS-6: Skip Mint On `cache.liquidityMinted == 0` After Resize (#83)

* skip minting if liquidity minted zero

* samples testing

* should skip mint if liquidityminted is zero

* LMP-6: Typo (#85)

* fix canonical typo

* fix canonical typo

* fix comments

* fix comments

* TK-9: Correct Value Emitted for `cache.liquidity` (#84)

* update cache.liquidity in _cross

* _pass consistent with _cross

* superflous claimTickEpoch updates (#86)

* superfluous casting (#91)

* RNP-1: Move Liquidity Overflow Checks (#88)

* liquidity overflow checks

* test for liquidity overflow

* round back tick before clearning position (#87)

* GLOBAL-11: Redundant Boolean Logic In _empty (#90)

* superfluous boolean setting

* range tick empty

* samples testing

* GLOBAL-12 - Clones Can Receive ETH (#89)

* withdraw eth safely using fees call

* test errantly sending eth to pool or pool token

* cleanup position tokens

* add tests for no position and owner mismatch

* MRC-1: Tick Validation Before Position Update (#94)

* liquidity overflow test; check ticks at start of mint

* liquidity overflow check

+++ contracts/base/events/LimitPoolFactoryEvents.sol
@@ -9,6 +9,6 @@ abstract contract LimitPoolFactoryEvents {
         address indexed token1,
         uint16 indexed swapFee,
         int16 tickSpacing,
-        uint8 poolTypeId
+        uint16 poolTypeId
     );
 }
